requirements.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.7
get-cover-art==1.3.11
mutagen==1.45.1
numpy==1.20.1
Pillow==8.1.2
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
